parsers: bring back parse duration logic into mpd parser
